test(ast/estree): ignore TS-ESTree tests where fail to parse (#10661)
Related to #9705.

Skip tests in TS-ESTree conformance which do fail to parse, but fail because of bugs which are not related to ESTree serialization.

We should eventually try to make these tests pass, but when we do, that'll show up in the parser conformance snapshots. So no need to repeat them in the ESTree conformance snapshot too - so we can more easily see what our ESTree-related problems are.

Note: `typescript/tests/cases/conformance/jsx/jsxReactTestSuite.tsx` is *not* included in ignore list because that's not a fail in parser conformance (see #10578).