test(ast/estree): skip test for `\r` in template literal (#10677)
Skip TS-ESTree conformance test for `\r` in `raw` field of `TemplateElement`. `@typescript-eslint/parser` is wrong on this one, and we can't massage the AST to match them.

We do the `\r` to `\n` conversion in parser, so the information that there were originally `\r` characters in the raw string is lost by the time we get to serialization.

We *could* get the slice of source text from the `Span`, and use that, but that only works if the AST has come direct from parser. If the AST has been through any form of transformation, the content of the `TemplateElement` could have been changed, and it'd be out of sync with the chunk of source text described by the `Span`. This seems like a recipe for bugs.