fix(playground): playground show ESTree AST for JS files (#10673)
The ESTree AST (for JS files) and TS-ESTree AST (for TS files) are quite different. Display the relevant AST in Playground, depending on source type.

Note: This will likely increase binary size of the WASM file, due to including 2 serializers. But in my opinion, it's worth it. Most people using the Playground will be developers who tend to have pretty fast internet connections.