refactor(ast)!: `AstBuilder` methods require an `Atom` with correct lifetime (#10735)
Closes #10734.

`AstBuilder` methods previously took `IntoIn<'a, Atom<'a>>`. This is a footgun because it makes it really easy to end up copying string data from an existing `Atom`, when that `Atom` could be reused for free. Just passing an `&Atom` instead of `Atom` is enough to accidentally trigger copying the whole string, incurring higher memory usage.

Instead, take `Into<Atom<'a>>`. This still allows the convenience of passing `&'static str` or `&'a str`, but these don't result in the string being copied, only referenced.

Additionally, only take a `&'a str` in `AstBuilder::program`, instead of `IntoIn<'a, &'a str>`, for the same reasons.

All the changes in this PR outside of `AstBuilder` itself are just to manually allocate `Atom`s in the few places where the string does not already exist in the arena. All the places where we were unnecessarily copying string data are already addressed in PRs that precede this one e.g. #10726.

The downside of this change is that it requires more verbose code in places, but I think that's outweighed by the benefit. Now when calling an `AstBuilder` method, you have to ask yourself "do I need to copy this string?" and you have to insert an `ast.atom(...)` call if you conclude that you need to. Previously you didn't have to consider that question, and poor performance/memory usage was the default.