perf(allocator/vec2): calling `Bump::grow` or `Bump::shrink` at the call site directly instead of calling `realloc` (#10686)
This PR aims to remove the `realloc` method, which was copied from `bumpalo`. The `realloc` function is used to call the `shrink` or `grow` method of bumpalo by checking the new size. This is unnecessary because in the call site, we know which method we should call, and the ZST check is also unnecessary because it has been checked earlier in the call site.

Also, this is aligned with the standard library implementation.