refactor(language_server): move `nested_configs` to `ServerLinter` (#10775)
Normally we would just restart the `ServerLinter.isolated_linter` or the `Worker.nested_configs`.
I do not see the benefits to store the `nested_configs` separately and changing them + the `isolated_linter` when needed.
And I do not want to change `isolated_linter` from `Arc` to `RwLock` / `Mutex`.

When a `.oxlintrc.json` file changes, we just set up the linter and searching for all nested configs again.