fix(linter): false positive with negative matches in no-restricted-imports (#10976)
this was an absolute nightmare to fix.
TLDR is that `GitignoreBuilder` nor `OverrideBuilder` do not contain the flexibility or have the behaviour we want to achieve this here. They don't go through the full list of patterns (for perf reasons), but that has the caveat of meaning it does not work correctly with eslint

closes #10911