perf(parser): Explore packed tokens (#10933)
Pack the parser tokens into a single u128.

I was musing with o3 about compiler performance, and it told me that
Hermes used packed tokens to improve parser performance so I thought I'd
try and see if that would also work here. In local benchmarks it appears
to have a decent improvement.

closes #9926