fix(linter): fix rule config not being correctly applied (#11055)
since `rules_for_override` was converted from a `FxHashSet` to ` FxHashMap`, the behaviour changed. rather than calling `.replace`, we call `insert` which updates the value, but does not update the key.

As per the docs:

> The key is not updated, though; this matters for types that can be == without being identical.

I've written a test for this to hopefully prevent a regression and validated the test fails on main.

fixes #11054