fix(napi/playground): Always display syntax errors (#11065)
Related #11029

Currently, if the `Check Syntax Errors` option is OFF, it appears that
no errors are found during parsing.


![image](https://github.com/user-attachments/assets/95c7c82e-0b53-4c3e-8327-54e73c5981ea)

But actually the parser returns `errors`, this is the gap between
playground and parser.


![image](https://github.com/user-attachments/assets/268d6895-6bd0-4744-89d6-bb848da13653)

---

Does this change make sense?
If so, I will update the playground label from `Check Syntax Errors` to
`Check Semantic Errors` later.