perf(allocator/vec): remove `alloc_guard` from `RawVec::from_raw_parts_in` + clarify safety docs (#11073)
Follow-on after #10884.

We don't need `alloc_guard` here. `RawVec::from_raw_parts_in` is an unsafe method, and caller guarantees that `len` and `capacity` are in legal range. We don't need to check the caller fulfilled that guarantee - that responsibility is on them.

Also clarify the safety docs for this method.