ci(miri): fix Miri (#11092)
Miri is failing on main. It seems to have started with #11088, though
that could be a coincidence - maybe it's just latest nightly made some
change which broke Miri.

2 things are causing the fails:

1. "fatal error: cross-interpreting doctests is not currently supported
by Miri".
2. A lint error that only occurs on nightly in `Vec::append_elements`.

Fix (1) by disabling doc tests.

Fix (2) by changing the code. The `#[allow]` attribute is needed because
the amended code triggers a *different* clippy warning on stable!