perf(ast): use bitflags for storing comment newline state (#11096)
- supersedes https://github.com/oxc-project/oxc/pull/11090

This switches the newline flags to using `bitflags` to make the setting of the flags a little bit simpler and also makes checking the flag state easier too. Rather than using all flags in a bitflags, this is simpler because we don't need a custom serializer for comments for ESTree.