feat(data_structures): introduce `PointerExt` trait (#11095)
Rust v1.87.0 introduced new methods for pointers, notably `offset_from_unsigned` which calculates the distance between 2 pointers.

https://doc.rust-lang.org/std/primitive.pointer.html#method.offset_from_unsigned

`offset_from_unsigned` is the same as `offset_from`, except it require the source pointer to be before the target, and so returns a `usize` rather than an `isize`. In return for this restriction, the compiler is able to generate more efficient code in some cases.

We want to use these methods, but our MSRV won't hit 1.87.0 for 3 months.

This PR introduces a `PointerExt` trait which provides similar methods `offset_from_usize` and `byte_offset_from_usize`.

On versions of Rust >= 1.87.0, these delegate directly to the native Rust methods `offset_from_unsigned` and `byte_offset_from_unsigned`. On earlier versions of Rust, they use a fallback implementation.

This PR replaces all use of `offset_from` with `offset_from_usize` in our crates.

This will probably make little difference to perf, as we don't use these methods much, and the perf advantage is small anyway. But reason I'm making this PR now is that allocator will make liberal use of them on very hot paths.