perf(mangler)!: `Mangler::build_with_semantic` take mut ref to `Semantic` (#11132)
`Semantic` is a very large type (600 bytes). The mangler only uses a part of it (`Scoping` and `AstNodes`), so it's a waste to copy the whole type onto the stack and then return it again.

Change `Mangler::build_with_semantic` to take `&mut Semantic` instead.

I hope this may also help a bit with wild variance in mangler benchmarks, as `Semantic` will no longer get dropped in the measured section (I don't think should be part of the measurement anyway, as the drop time of `Semantic` is not relevant in this benchmark).