refactor(allocator/vec): access `len` and `cap` fields via getters/setters (#11081)
Follow-on after #10884.

Pure refactor.

To make it explicit when we're converting `u32` to/from `usize`:

* Make `RawVec`'s `len` field private.
* Make all access to `len` and `cap` fields go via getter/setter methods.
* Name those methods `*_u32` and `*_usize` so it's clear when a conversion is happening.

I also renamed `cap_*` methods to `capacity_*` for consistency with `Vec::capacity`.

I haven't used these methods in `RawVec` because there's some subtlety around when you want the "raw" `cap` value, and when you want the value `capacity` method returns, which is different when `T` is a ZST. Hopefully we can remove support for ZSTs, and remove this ambiguity.