fix(allocator/vec): fix unsoundness in `Vec::extend_from_slices_copy` (#11200)
Fix a soundness bug in `Vec::extend_from_slices_copy`.

Similar to `String::from_strs_array_in`, we have to handle the possibility that the total length of the slices provided could exceed `usize::MAX`. If it did, we could reserve too little capacity, and write out of bounds. Used checked addition to catch this.

This bug is present in `bumpalo`, and we inherited it.