refactor(allocator): introduce `Alloc` trait (#11198)
First step towards replacing `bumpalo` with our own allocator (#11145).

Introduce `Alloc` trait and implement it for `bumpalo::Bump`. A later PR will parameterize `Vec` with `Alloc`, instead of a static dependency on `Bump`.

We can later on replace `Bump` with our own `Arena` type. `Arena` will also implement `Alloc`, so the two will be able to be swapped with a minimum of fuss.

### Why do we need a trait at all?

Why couldn't `Vec` just statically depend on `Arena`, same as it now does on `Bump`?

Rationale for a trait here is: My intent is to have 2 different variants of `Arena` to support filling allocator chunks from both ends (string data fills from the start, all other types fill from the end). This will have 3 advantages:

1. More tightly packed data. `&str` / `Atom` is the only type in the AST which is not aligned on 8. Strings currently often tend to end up with unnecessary padding bytes around them.
2. Packing all string data into a single immutable stretch of memory has potential to speed up all string operations (https://github.com/oxc-project/backlog/issues/46#issuecomment-2603153807).
3. It should unblock the biggest current perf bottleneck in raw transfer, by allowing all strings in the AST to be converted to UTF-16 in a single shot (as if they were just 1 long string), rather than converting each string individually.

`String` is a wrapper around `Vec`, and it's preferable not to have to implement `Vec` twice for front-filling (strings) and back-filling (all other types). `Vec<T, A: Alloc>` is a simpler solution.