feat(allocator): introduce `StringBuilder` (#11257)
First step towards #11246.

Introduce `StringBuilder` type for constructing a string in the arena.

Difference from `String` is that it's optimized for building strings, not querying them etc.

Should only be used as a temporary value while constructing a string, and then converted to an immutable `&'a str` with the lifetime of the arena, using `StringBuilder::into_str`. It is *not* designed for storing permanently (it's 8 bytes bigger than `String`), and not designed for efficient string comparison (a little less efficient).

This pattern is exactly what we currently use `String` for. The AST and `Semantic` only contains `&str`s and `Atom`'s, and we only use `String` as an intermediary to construct those `&str`s. So `StringBuilder` is more optimal for our use case than `String`.

That said, we don't use `String` much anyway, so it's unlikely to move the needle on benchmarks.