perf(linter/no-unused-vars): simplify checking self call usage (#11281)
Use `Span::contains_inclusive` to simplify checking self call.

When the reference is inside a function or class declaration (e.g., `function foo() { foo() }` or `class Foo { bar() { new Foo() } }`), then the span of the declaration will contain the span of the reference for sure.

In addition, I've added some comments to describe the purpose of this function.