refactor(parser): align jsx parsing to tsc (#11314)
My original plan was to rework the JSX parser from the entry point, as I had found in #11232.

However, when I read through all the TSC JSX parser code, I found that what we are doing is not much different.

---

As for perf perspective, the OXC currently requires rewind in 3 places for JSX:

- 1: `parse_jsx_expression()` - `lookahead`
  - To determine if it is a `JSXElement`(`<C`) OR `JSXFragment`(`<>`) at the entry point
- 2: `parse_jsx_child()` - `checkpoint+rewind`
  - The same above OR breaking list(`</`)
- 3: `parse_jsx_child()` - `checkpoint+rewind`
  - To determine if it is `JSXChild::Spread`(`{...Expression}`) OR `JSXChild::ExpressionContainer`(`{JSXExpression}`)

These don't exist in TSC, they just use union type or loose ASTs.

But we have different and explicit ASTs in Rust, it seems they can't be easily eliminated.

And unfortunately, I tried, but there was no remarkable improvement.

- For 1, 3045952596c7201348b26ef347bdd99e110384e4 => Minifier +3%, but no effect for parser
- For 2, I wish we had `Kind::LessThanSlash`...
- For 3, a12ff218e6d93be84a90ba1f8bcff25181b4b384 => No effect for parser

As a result, I think it would be better to keep the current clean and explicit code.