fix(linter)!: remove react from default plugin set (#11382)
This is a breaking change, but currently, how we check whether a react rule should run is opaque and unclear.

For example, the `react/forward-ref-uses-ref` rule checks whether the framework flags `FrameworkFlags::React` is enabled, however this is never set to true meaning that this rule can never run.

https://github.com/oxc-project/oxc/blob/bbb7eb1a27ac280febc3702bc5e7eb131dcc005a/crates/oxc_linter/src/rules/react/forward_ref_uses_ref.rs#L99-L101

This PR removes the react rule group from the default plugins, if users which to use the react plugin, they just have to add it to the plugins array.

I have made thie change before 1.0, so hopefully the fact that it's breaking is ok.