feat(prettier): fill printing `CallExpression` implementation (#11299)
Printing a call expression correctly is mostly done, except for printing comments. Not many tests have passed because there is a problem with `ParentStack`. Some places use `ParentStack` to check the current parent, which may be incorrect. The issue is that some printing implementations do not call `push`/`pop`, meaning the recorded parent node is not accurate—it only reflects the last pushed record. I’ve attempted to solve this problem in #11339.