fix(linter/no-inner-declarations): flag `var` statement as body of `for` loop (#11632)
Follow-on after #11617. That change broke the rule (failed to report errors) in these cases:

```js
for (const x in {}) var y = 5;
for (const x of []) var y = 5;
for (const x = 1; a < 10; a++) var y = 5;
```

We need to check that parent is a `for` statement *and* that the declaration is its `init` / `left`.